Update collisionsExample
I'm putting this up so you can see the processing code. Ultimately, I don't know if "keypressed" exists in C++, but since I know you were looking into us doing that. Also, I think possibly having a separate "edgeCollide" like this code does might be able to clear any of the glitches it is experiencing? Lastly, I think ballCollide is the most important here. I don't know where all of the references are to where things are being called, but I'll look at it more later. I just wanted you to be able to see it also. (this code doesn't have the best labeling or notes to be able to understand it)